Add recursion guard to Analyzer::parse_expr
